requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
reportlab==4.0.7
weasyprint==60.2
django-import-export==3.3.1
drf-spectacular==0.27.0
django-filter==23.5
django-extensions==3.2.3
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0
django-storages==1.14.2
boto3==1.34.0
Pillow==10.2.0
